Wire: connectionId is a string, and a parse failure names the field
Third of the three "the suite certified the wrong behaviour" defects from
the TestRig port, plus a second instance of it found by auditing the rest
of the assembly, plus the reason both were invisible.

ConnectedClient.ConnectionId was int? against a long RakNet connection id.
Measured values: 189151461494586169 and 1044835390751713754. System.Text.Json
threw on the value, RigWire.Deserialize caught JsonException and returned null
by design, and the ENTIRE /status payload vanished rather than one field. The
host's roster read as empty and four of eight playtest checks reported
inconclusive (joiner-not-in-roster) against a rig that was joining perfectly.
WireTrapFieldTests asserted ConnectionId == 2, which is why 1,688 tests missed it.

It is a string now, not a long, matching the clientId beside it in the same
row. Both are identities past 2^53, neither is ever computed with, and two ids
side by side spelled two different ways is the trap the next reader falls into.
The plugin therefore stops degrading: it emitted the number only when it
round-tripped Int32 and always supplied connectionIdString beside it, and both
that condition and that second spelling are gone. Nothing read either; the only
reader of ConnectionId anywhere was the unit test.

Auditing every other field in Contracts against the plugin emitter found one
more live case and one latent one:

- ConsoleBufferLine.Color was int against ConsoleWindow.ConsoleLine.Color,
  a uint packed ImGui colour. The alpha byte sets the high bit, so the DEFAULT
  colour is past int.MaxValue and one row was enough to take the whole
  /console/buffer response down the same way. Now uint. The old sample used
  the value 3, exactly as the connectionId sample used 2.
- DlcState.BaselineSession and DlcRestoreResponse.BaselineSession were int?
  while the value copied into them is Epoch.Session, a long, and
  EpochBlock.Session is already long. Never near either bound, but a wire type
  narrower than its source is the whole defect class. Now long?.

RigWire.Deserialize no longer turns a loud, specific failure into a silent,
total one. Null now means only "the plugin sent nothing"; a body that is
present and does not fit throws RigWireFormatException naming the field, what
the contract wanted and what arrived ("connectedClients[0].connectionId ...
the number 189151461494586169"). A throw rather than a quieter null because no
caller can act on it: a body that does not fit the contract fits no better on
the next poll, so the reader loops that treated null as "not ready yet" burned
their whole timeout and then blamed the game. The runner classifies it as
inconclusive/wire-format, so it accuses the wire and never the mod.

Suite is 1,739, up 11.

+++ TestRig/src/TestRig.Contracts/Dlc/DlcEndpoints.cs
@@ -93,9 +93,19 @@ public sealed record DlcState
     [JsonPropertyName("removedShared")]
     public string? RemovedShared { get; init; }
 
-    /// <summary>The <c>epoch.session</c> the baseline was captured in. A later session invalidates it.</summary>
+    /// <summary>
+    ///     The <c>epoch.session</c> the baseline was captured in. A later session
+    ///     invalidates it.
+    /// </summary>
+    /// <remarks>
+    ///     A <c>long</c>, matching <see cref="EpochBlock.Session"/>, which is the value the
+    ///     plugin copies in here. It was <c>int?</c>, so one field said the session counter
+    ///     was 32 bits wide and another said 64. The counter never gets near either bound,
+    ///     but a wire type narrower than its source is the defect class that made
+    ///     <c>connectionId</c> take a whole endpoint down, and agreeing costs nothing.
+    /// </remarks>
     [JsonPropertyName("baselineSession")]
-    public int? BaselineSession { get; init; }
+    public long? BaselineSession { get; init; }
 
     [JsonPropertyName("removeCalls")]
     public int? RemoveCalls { get; init; }
@@ -306,8 +316,9 @@ public sealed record DlcRestoreResponse : IWireResult
     [JsonPropertyName("sharedAfter")]
     public string? SharedAfter { get; init; }
 
+    /// <summary>A <c>long</c>, matching <see cref="EpochBlock.Session"/> and <see cref="DlcState.BaselineSession"/>.</summary>
     [JsonPropertyName("baselineSession")]
-    public int? BaselineSession { get; init; }
+    public long? BaselineSession { get; init; }
 
     [JsonPropertyName("error")]
     public string? Error { get; init; }
